Update main.cpp
created the main two parts (testing roulette and the gambler), but the gambler is flawed. The code generates his numbers and the spins separately and then determines how many times he won. The logic should instead test the win during the spin, i.e. the roulette spins and the gambler guesses a number, if he guessed right it's a win otherwise it's a loss.